- renamed  answer  to quakifiedPersons - deleted the events class - renamed EventsManager to Demo - moved the functions that call the events to the Porgramm class
