+++ Src/BootCamp.Chapter/Demo.cs
@@ -1,11 +1,11 @@
 ﻿using System;
-using System.Collections.Generic;
-using System.Text;
 
 namespace BootCamp.Chapter
 {
-    public class Event
+    public class Demo
     {
+        public Person person;
+
         public event EventHandler DemoStarted;
 
         public event EventHandler ChoosePredicate;
@@ -14,13 +14,12 @@ namespace BootCamp.Chapter
 
         public event EventHandler ApplicationClosed;
 
-        internal void Run()
+        public void Run()
         {
             DemoStarted?.Invoke(this, EventArgs.Empty);
             ChoosePredicate?.Invoke(this, EventArgs.Empty);
             DemoEnded?.Invoke(this, EventArgs.Empty);
             ApplicationClosed?.Invoke(this, EventArgs.Empty);
         }
-        
     }
 }
